wip(ui-design-tool): remove unnecessary code on SelectionOverlay
